Deploy the test app into the specified app domain
When pinging, we use the app-name.app-domain path, however we weren't
specifying that the app was pushed into 'app-domain'. As a result, this
only passes when the first domain returned by the cf-cli is the same as
'app-domain'

Also we added pikachu as its our current default testing environment and
a readme to explain usage.

[#152430627]

Signed-off-by: Jack Newberry <jnewberry@pivotal.io>