Revert "Fix for RFPAS-29"
Smoke tests used to use cf-cliv6
$ cf disable-service-access -o system p.redis-ci -p cache-small
Disabling access to plan cache-small of service p.redis-ci for org system as admin...
OK

Would yield no errors for public plans, but on cf-cliv8
$ cf disable-service-access -o system p.redis-ci -p cache-small
Disabling access to plan cache-small of service offering p.redis-ci for org system as admin...
Cannot remove organization level access for public plans.
FAILED

This is why the bug was introduced, we first disable-service-access, as
the docs suggest, then enable it for all orgs.

What we want is just to ignore public plans, as they do not need to be
tested on enable service access

This reverts commit bb4b1e37b0abcec6615498eb691e52cc6ef6db19.